+++ src/app/app.component.html
@@ -219,77 +219,3 @@
     </div>
   </div>
 </div>
-
-<!-- Server Dialog -->
-<div class="modal fade" id="serverDialog" role="dialog" aria-hidden="true" *ngIf="isAdmin">
-  <div class="modal-dialog modal-dialog-centered user-modal" role="document">
-    <div mdcDialogSurface class="modal-content border-0">
-      <mdc-dialog-title>Server Details</mdc-dialog-title>
-      <mdc-dialog-content>
-        <mdc-list disableRipple [interactive]="false" border>
-          <mdc-list-item style="height: auto">
-            <mdc-icon mdcListItemGraphic svgIcon="angularjs"></mdc-icon>
-            <p class="p-0 m-0">
-              <b>WebApp Build Time:</b>&nbsp;{{ server.build_time || (server.is_dev ? 'Development' : 'Loading') }}<br/>
-              <b>WebApp Last Commit:</b>&nbsp;{{ server.is_dev ? 'Development' : '' }}
-              <a href="https://github.com/orange-alliance/the-orange-alliance/commit/{{ server.last_commit }}" target="_blank" *ngIf="!server.is_dev">
-                <code>{{ server.last_commit ? server.last_commit.substr(0, 7) : 'Error' }}</code>
-              </a>
-              <a *ngIf="serverData" class="mdc-list-item__secondary-text">Uptime: {{calculateUptime(serverData[1].pm2_env.pm_uptime)}}</a>
-            </p>
-            <a mdcListItemMeta *ngIf="isToaDev" mdcIcon title="Restart" (click)="updateService(services.Live)">refresh</a>
-          </mdc-list-item>
-          <mdc-list-item>
-            <mdc-icon mdcListItemGraphic svgIcon="webhook"></mdc-icon>
-            <b>API Version:</b>&nbsp;{{ server.api_version || 'Loading' }}
-            <a mdcListItemMeta *ngIf="isToaDev"  mdcIcon title="Restart" (click)="updateService(services.Api)">refresh</a>
-          </mdc-list-item>
-          <mdc-list-item>
-            <mdc-icon mdcListItemGraphic svgIcon="test-tube"></mdc-icon>
-            <b>Update Dev WebApp</b>
-            <a mdcListItemMeta *ngIf="isToaDev"  mdcIcon title="Restart" (click)="updateService(services.Dev)">refresh</a>
-          </mdc-list-item>
-          <mdc-list-item>
-            <mdc-icon mdcListItemGraphic svgIcon="material-design"></mdc-icon>
-            <b>MDC Version:</b>&nbsp;{{ server.mdc_version }}
-          </mdc-list-item>
-          <mdc-list-item style="height: auto">
-            <mdc-icon mdcListItemGraphic svgIcon="update"></mdc-icon>
-            <p class="p-0 m-0">
-              <b>Auto Deploy</b>
-              <a *ngIf="serverData" class="mdc-list-item__secondary-text">Uptime: {{calculateUptime(serverData[2].pm2_env.pm_uptime)}}</a>
-            </p>
-            <a *ngIf="serverData && serverData[2].pm2_env.status === 'online'" mdcListItemMeta mdcIcon style="color: green" title="Online">done</a>
-            <a *ngIf="serverData && serverData[2].pm2_env.status === 'stopped'" mdcListItemMeta mdcIcon style="color: red" title="Stoped">stop</a>
-            <a *ngIf="serverData && serverData[2].pm2_env.status === 'errored'" mdcListItemMeta mdcIcon style="color: red" title="Errored">error</a>
-          </mdc-list-item>
-          <mdc-list-item style="height: auto">
-            <mdc-icon mdcListItemGraphic>textsms</mdc-icon>
-            <p class="p-0 m-0">
-              <b>TOA Text</b>
-              <a *ngIf="serverData" class="mdc-list-item__secondary-text">Uptime: {{calculateUptime(serverData[3].pm2_env.pm_uptime)}}</a>
-            </p>
-            <a *ngIf="serverData && serverData[3].pm2_env.status === 'online'" mdcListItemMeta mdcIcon style="color: green" title="Online">done</a>
-            <a *ngIf="serverData && serverData[3].pm2_env.status === 'stopped'" mdcListItemMeta mdcIcon style="color: red" title="Stoped">stop</a>
-            <a *ngIf="serverData && serverData[3].pm2_env.status === 'errored'" mdcListItemMeta mdcIcon style="color: red" title="Errored">error</a>
-          </mdc-list-item>
-          <!--
-          <mdc-list-item style="height: auto">
-            <mdc-icon mdcListItemGraphic>save</mdc-icon>
-            <p class="p-0 m-0">
-              <b>Redis Cache</b>
-              <a *ngIf="serverData" class="mdc-list-item__secondary-text">Uptime: {{calculateUptime(serverData[4].pm2_env.pm_uptime)}}</a>
-            </p>
-            <a *ngIf="serverData && serverData[4].pm2_env.status === 'online'" mdcListItemMeta mdcIcon style="color: green" title="Online">done</a>
-            <a *ngIf="serverData && serverData[4].pm2_env.status === 'stopped'" mdcListItemMeta mdcIcon style="color: red" title="Stoped">stop</a>
-            <a *ngIf="serverData && serverData[4].pm2_env.status === 'errored'" mdcListItemMeta mdcIcon style="color: red" title="Errored">error</a>
-          </mdc-list-item>
-          -->
-        </mdc-list>
-      </mdc-dialog-content>
-      <mdc-dialog-actions>
-        <button mdcDialogButton data-dismiss="modal">Close</button>
-      </mdc-dialog-actions>
-    </div>
-  </div>
-</div>
