+++ src/app/components/server/server.item.component.scss
@@ -1,0 +1,21 @@
+@import "../../../constants";
+
+[dir="ltr"].data-indicator {
+  background: $mdc-theme-primary;
+  top: 4px;
+  left: 4px;
+  width: 4px;
+  bottom: 4px;
+  position: absolute;
+  border-radius: 4px;
+}
+
+[dir="rtl"].data-indicator {
+  background: $mdc-theme-primary;
+  top: 4px;
+  right: 4px;
+  width: 4px;
+  bottom: 4px;
+  position: absolute;
+  border-radius: 4px;
+}
