+++ src/app/views/account/account.component.html
@@ -8,11 +8,11 @@
   </div>
 </div>
 
-<div class="full-page" *ngIf="user">
+<div class="full-page" *ngIf="user && user.firebaseUser">
   <div class="profile-block px-3 px-lg-5 pt-3 pb-5">
     <div *ngIf="!user.photoURL && user.displayName" class="profile-image">{{ user.displayName.substring(0, 1) }}</div>
     <img *ngIf="user.photoURL " class="profile-image" [src]="user.photoURL" />
-    <div class="d-flex flex-column w-100 ps-3">
+    <div class="flex-column w-100 ps-3">
       <div class="d-flex justify-content-between">
         <div class="profile-name">{{ user.displayName || 'User' }}</div>
         <button mdc-button primary (click)="signOut()">{{ 'pages.account.logout' | translate }}</button>
@@ -57,6 +57,10 @@
         </div>
       </div>
       <div class="col-md-5 col-12 ps-md-3">
+        <div class="account-card p-3 mb-3">
+          <div mdcHeadline5 class="black mt-2">Server</div>
+          <toa-server [user]="user"></toa-server>
+        </div>
         <div class="account-card space-between p-3 mb-3" *ngIf="notifications && isSupported">
           <div class="icon-panel">
             <div mdcHeadline5 class="black mt-2">{{ 'pages.account.notifications_card.title' | translate }}</div>
